moving loading to middleware
