+++ x-pack/plugins/lens/public/state_management/index.ts
@@ -5,16 +5,15 @@
  * 2.0.
  */
 
-import { configureStore, DeepPartial, getDefaultMiddleware } from '@reduxjs/toolkit';
+import { configureStore, getDefaultMiddleware, DeepPartial } from '@reduxjs/toolkit';
 import logger from 'redux-logger';
 import { useDispatch, useSelector, TypedUseSelectorHook } from 'react-redux';
 import { lensSlice, initialState } from './lens_slice';
 import { timeRangeMiddleware } from './time_range_middleware';
 import { optimizingMiddleware } from './optimizing_middleware';
-import { externalContextMiddleware } from './external_context_middleware';
-
-import { DataPublicPluginStart } from '../../../../../src/plugins/data/public';
-import { LensAppState, LensState } from './types';
+import { LensAppState, LensState, StoreDeps } from './types';
+import { getInitialDatasourceId, getResolvedDateRange } from '../utils';
+import { initMiddleware } from './init_middleware';
 export * from './types';
 
 export const reducer = {
@@ -22,8 +21,9 @@ export const reducer = {
 };
 
 export const {
-  setState,
+  loadInitial,
   navigateAway,
+  setState,
   setSaveable,
   onActiveDataChange,
   updateState,
@@ -38,36 +38,65 @@ export const {
   setToggleFullscreen,
 } = lensSlice.actions;
 
-export const getPreloadedState = (initializedState: Partial<LensAppState>) => {
+export const getPreloadedState = (
+  {
+    lensServices: { data },
+    initialContext,
+    embeddableEditorIncomingState,
+    datasourceMap,
+    visualizationMap,
+  }: StoreDeps,
+  preloadedState: Partial<LensAppState>
+) => {
+  const initialDatasourceId = getInitialDatasourceId(datasourceMap);
+  const datasourceStates: LensAppState['datasourceStates'] = {};
+  if (initialDatasourceId) {
+    datasourceStates[initialDatasourceId] = {
+      state: null,
+      isLoading: true,
+    };
+  }
+
   const state = {
     lens: {
       ...initialState,
-      ...initializedState,
+      isLoading: true,
+      query: data.query.queryString.getQuery(),
+      // Do not use app-specific filters from previous app,
+      // only if Lens was opened with the intention to visualize a field (e.g. coming from Discover)
+      filters: !initialContext
+        ? data.query.filterManager.getGlobalFilters()
+        : data.query.filterManager.getFilters(),
+      searchSessionId: data.search.session.getSessionId(),
+      resolvedDateRange: getResolvedDateRange(data.query.timefilter.timefilter),
+      isLinkedToOriginatingApp: Boolean(embeddableEditorIncomingState?.originatingApp),
+      activeDatasourceId: initialDatasourceId,
+      datasourceStates,
+      visualization: {
+        state: null as unknown,
+        activeId: Object.keys(visualizationMap)[0] || null,
+      },
+      ...preloadedState,
     },
   } as DeepPartial<LensState>;
   return state;
 };
 
-type PreloadedState = ReturnType<typeof getPreloadedState>;
-
-export const makeConfigureStore = (
-  preloadedState: PreloadedState,
-  { data }: { data: DataPublicPluginStart }
-) => {
+export const makeConfigureStore = (storeDeps: StoreDeps, preloadedState: Partial<LensAppState>) => {
   const middleware = [
     ...getDefaultMiddleware({
       serializableCheck: false,
     }),
+    initMiddleware(storeDeps),
     optimizingMiddleware(),
-    timeRangeMiddleware(data),
-    externalContextMiddleware(data),
+    timeRangeMiddleware(storeDeps.lensServices.data),
   ];
   if (process.env.NODE_ENV === 'development') middleware.push(logger);
 
   return configureStore({
     reducer,
     middleware,
-    preloadedState,
+    preloadedState: getPreloadedState(storeDeps, preloadedState),
   });
 };
